+++ packages/manager/apps/hub/src/billing/components/billing-status/BillingStatus.component.tsx
@@ -48,6 +48,7 @@ export default function BillingStatus({ service }: BillingStatusProps) {
           size={ODS_CHIP_SIZE.sm}
           color={ODS_THEME_COLOR_INTENT[BADGE_INTENT_BY_STATUS[service.getRenew()]]}
           inline
+          className="min-h-min"
         >
           <OsdsText
             color={ODS_THEME_COLOR_INTENT.text}