+++ packages/manager-ui-kit/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/ovh/manager/compare/@ovh-ux/muk@0.1.0...@ovh-ux/muk@0.1.1) (2025-10-23)
+
+**Note:** Version bump only for package @ovh-ux/muk
+
 # [0.1.0](https://github.com/ovh/manager/compare/@ovh-ux/muk@0.0.1...@ovh-ux/muk@0.1.0) (2025-10-23)
 
 ### Bug Fixes