+++ packages/manager/apps/container/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.161.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-container-app@0.161.2...@ovh-ux/manager-container-app@0.161.3) (2025-10-28)
+
+**Note:** Version bump only for package @ovh-ux/manager-container-app
+
+
+
+
+
+## [0.161.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-container-app@0.161.1...@ovh-ux/manager-container-app@0.161.2) (2025-10-27)
+
+**Note:** Version bump only for package @ovh-ux/manager-container-app
+
+
+
+
+
 ## [0.161.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-container-app@0.161.0...@ovh-ux/manager-container-app@0.161.1) (2025-10-23)
 
 **Note:** Version bump only for package @ovh-ux/manager-container-app